Saving work; updated documentation; performing final validation
