+++ setup.cfg
@@ -2,7 +2,7 @@
 name = fpds
 author = Derek Herincx
 author_email = derek663@gmail.com
-version = 1.1.0
+version = 1.2.0
 description = A parser for the Federal Procurement Data System (FPDS) Atom feed
 long_description = file: README.md
 long_description_content_type = text/markdown
